refactor: removed params file
