+++ jarust/src/jasession.rs
@@ -1,5 +1,5 @@
 use crate::jahandle::JaHandle;
-use crate::params::AttachHandleParams;
+use crate::japlugin::AttachHandleParams;
 use crate::prelude::*;
 use async_trait::async_trait;
 use jarust_rt::JaTask;
