+++ jarust_plugins/examples/video_room.rs
@@ -1,13 +1,12 @@
-use std::path::Path;
-
-use tracing_subscriber::EnvFilter;
-
-use jarust::jaconfig::{JaConfig, TransportType};
-use jarust::params::CreateConnectionParams;
+use jarust::jaconfig::JaConfig;
+use jarust::jaconfig::TransportType;
+use jarust::jaconnection::CreateConnectionParams;
 use jarust::TransactionGenerationStrategy;
 use jarust_plugins::video_room::jahandle_ext::VideoRoom;
 use jarust_plugins::video_room::msg_options::{VideoRoomAllowedAction, VideoRoomEditOptions};
 use jarust_plugins::AttachPluginParams;
+use std::path::Path;
+use tracing_subscriber::EnvFilter;
 
 #[tokio::main(flavor = "current_thread")]
 async fn main() -> anyhow::Result<()> {
