+++ jarust_plugins/src/audio_bridge/jahandle_ext.rs
@@ -1,7 +1,7 @@
 use super::events::PluginEvent;
 use super::handle::AudioBridgeHandle;
 use crate::AttachPluginParams;
-use jarust::params::AttachHandleParams;
+use jarust::japlugin::AttachHandleParams;
 use jarust::prelude::*;
 use std::ops::Deref;
 use tokio::sync::mpsc;
